truckersmp.py: Combine get_beta_branch_name() with determine_game_branch()
The new function name is determine_game_branch().
