+++ truckersmp_cli/steamcmd.py
@@ -15,30 +15,9 @@ import urllib.parse
 import urllib.request
 from zipfile import ZipFile
 
-from .truckersmp import get_beta_branch_name
+from .truckersmp import determine_game_branch
 from .utils import check_steam_process
 from .variables import Args, Dir, URL
-
-
-def determine_game_branch():
-    """
-    Determine Steam game branch name.
-
-    When "--beta" option is specified, this returns the specified branch.
-    Otherwise, this try to determine the branch using get_beta_branch_name():
-    If downgrade is needed, this returns the returned branch name.
-    If not needed, this returns the name "public" for using the latest version.
-    """
-    branch = "public"
-    if Args.beta:
-        branch = Args.beta
-    else:
-        game = "ats" if Args.ats else "ets2"
-        beta_branch_name = get_beta_branch_name(game)
-        if beta_branch_name:
-            branch = beta_branch_name
-
-    return branch
 
 
 def update_game():
